requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-functions==1.23.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.99.1
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2==3.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
